changed volume in order from double to int, restructured tests

+++ src/simulator/Trade.java
@@ -9,7 +9,7 @@ public class Trade extends Order {
 			String time,
 			String recordType,
 			double price,
-			double volume,
+			int volume,
 			String qualifiers,
 			long ID,
 			String bidAsk,
